+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -157,7 +157,7 @@ public class AddCommandTest {
         }
 
         @Override
-        public int totalModularCredits() {
+        public float totalModularCredits() {
             throw new AssertionError("This method should not be called.");
         }
